Orchestrator/requirements.txt
# Orchestrator - Minimal AI Agent System
# Focus: Reliability over features

# Core
crewai>=0.51.1
crewai-tools>=0.8.3
anthropic>=0.70.0
langchain-anthropic>=0.1.1
langchain-openai>=0.0.5

# UI
streamlit==1.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0

# File operations
pathlib2==2.3.7

# Database (audit logging)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver (optional, for production)

# Security
bandit==1.7.7
